+++ index.js
@@ -43,6 +43,7 @@ const safeObject = require('./util/safe_object');
  * @property {false} [youtube] Disable YouTube embeds.
  * @property {false} [wistia] Disable Wistia embeds.
  * @property {false} [twitter] Disable Twitter embeds.
+ * @property {false} [instagram] Disable Instagram embeds.
  * @property {false} [underline] Disable underline syntax.
  * @property {false|import('./modifiers/fence_label').FenceLabelOptions} [fence_label] Disable fence labels, or set options for the feature.
  * @property {false|import('./modifiers/fence_secondary_label').FenceSecondaryLabelOptions} [fence_secondary_label] Disable fence secondary labels, or set options for the feature.
@@ -142,6 +143,10 @@ module.exports = (md, options) => {
 
     if (optsObj.twitter !== false) {
         md.use(require('./rules/embeds/twitter'), safeObject(optsObj.twitter));
+    }
+
+    if (optsObj.instagram !== false) {
+        md.use(require('./rules/embeds/instagram'), safeObject(optsObj.instagram));
     }
 
     // Register modifiers